factory/simple_factory_test.go
package factory

import (
	"reflect"
	"testing"
)

func TestNewIRuleConfigParser(t *testing.T) {
	type args struct {
		t string
	}
	tests := []struct {
		name string
		args args
		want IRuleConfigParser
	}{
		{
			name: "json",
			args: args{t: "json"},
			want: jsonRuleConfigParser{},
		},
		{
			name: "yaml",
			args: args{t: "yaml"},
			want: yamlRuleConfigParser{},
		},
	}

	for _, test := range tests {
		t.Run(test.name, func(t *testing.T) {
			if got := NewIRuleConfigParser(test.args.t); !reflect.DeepEqual(got, test.want) {
				t.Errorf("NewIRuleConfigParser() = %v, want %v", got, test.want)
			}
		})
	}
}
